Create inches.c
The exercise prompt says to convert centimeters to inches or inches to centimeters, which is completed by line 19; however, I wanted to provide a foot and inches output (like 5ft 10 inches).

To do so, I took the total amount of inches and divided by 12 (70.08/12 = 5.84), then converted that value to an integer so I could be left with just the whole number value (5). From there, I used the float value, and subtracted it from the integer I just created to leave me with just the decimal value (5.84-5 = 0.84). This decimal value, of course, isn't the actual inches in a foot, so I multiply by 12 to get the actual inches rather than a fraction of a foot (0.84*12 = 10.08).

I'm sure there are better ways to do this, especially alternatives to scanf, converting a float to an int, and combining different data types. Based on what I've learned so far, this is what I've come up with. I'll be sure to add a revised version when I come across better ways to write this program.