Create plus_ten.c
This exercise didn't have an answer,  so I might have taken a different approach than others. The prompt was to create a program that took an integer provided by a user and then display that number along with the next ten numbers in sequence separating them by a space, tab, or new line.

The minimum solution was fairly simple: prompt for input, loop through the numbers starting at the input value and finishing at the value plus 10. However, I wanted to allow the option to do this as many times as the user wants, and allow an entry of '0' to quit the program. I did this using nested while loops.

I'm sure there are other ways, and perhaps better ways to do this, but I'm pretty happy with the result having only used what I've learned from the book so far. Of course, the obvious flaw is if the user inputs anything other than an integer, but I plan to update these as I cross that bridge.