Create 7.c
The solution to this exercise is available, and there are several differences from mine and the one provided which make my code significantly longer.

First is the use of comments which include the description, and I allow for a lot of white space between sections of code to increase readability.

The second thing would be the variable names I chose, which are a little longer than the ones in the answer guide. This forced me to break some lines into two so as to stay within 79 characters long.

Lastly, I added an extra if - else condition to filter out non-numerical entries.